break out categories creation
