+++ package.json
@@ -23,7 +23,7 @@
     "gatsby-plugin-sitemap": "^2.2.27",
     "gatsby-plugin-typography": "^2.3.21",
     "gatsby-source-filesystem": "^2.1.46",
-    "gatsby-source-graphql": "^2.1.25",
+    "gatsby-source-graphql": "^2.1.31",
     "gatsby-transformer-sharp": "^2.3.7",
     "gatsby-wpgraphql-inline-images": "^0.2.5",
     "html-react-parser": "^0.10.0",
